+++ sdev/include/shaiya/src/ItemHelper.cpp
@@ -8,7 +8,7 @@
 #include "include/shaiya/include/NetworkHelper.h"
 using namespace shaiya;
 
-void ItemHelper::SetCraftStrength(CItem* item, uint16_t value)
+void ItemHelper::SetCraftStrength(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -22,7 +22,7 @@ void ItemHelper::SetCraftStrength(CItem* item, uint16_t value)
     item->craftName[1] = str[1];
 }
 
-void ItemHelper::SetCraftDexterity(CItem* item, uint16_t value)
+void ItemHelper::SetCraftDexterity(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -36,7 +36,7 @@ void ItemHelper::SetCraftDexterity(CItem* item, uint16_t value)
     item->craftName[3] = str[1];
 }
 
-void ItemHelper::SetCraftReaction(CItem* item, uint16_t value)
+void ItemHelper::SetCraftReaction(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -50,7 +50,7 @@ void ItemHelper::SetCraftReaction(CItem* item, uint16_t value)
     item->craftName[5] = str[1];
 }
 
-void ItemHelper::SetCraftIntelligence(CItem* item, uint16_t value)
+void ItemHelper::SetCraftIntelligence(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -64,7 +64,7 @@ void ItemHelper::SetCraftIntelligence(CItem* item, uint16_t value)
     item->craftName[7] = str[1];
 }
 
-void ItemHelper::SetCraftWisdom(CItem* item, uint16_t value)
+void ItemHelper::SetCraftWisdom(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -78,7 +78,7 @@ void ItemHelper::SetCraftWisdom(CItem* item, uint16_t value)
     item->craftName[9] = str[1];
 }
 
-void ItemHelper::SetCraftLuck(CItem* item, uint16_t value)
+void ItemHelper::SetCraftLuck(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -92,7 +92,7 @@ void ItemHelper::SetCraftLuck(CItem* item, uint16_t value)
     item->craftName[11] = str[1];
 }
 
-void ItemHelper::SetCraftHealth(CItem* item, uint16_t value)
+void ItemHelper::SetCraftHealth(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -106,7 +106,7 @@ void ItemHelper::SetCraftHealth(CItem* item, uint16_t value)
     item->craftName[13] = str[1];
 }
 
-void ItemHelper::SetCraftMana(CItem* item, uint16_t value)
+void ItemHelper::SetCraftMana(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -120,7 +120,7 @@ void ItemHelper::SetCraftMana(CItem* item, uint16_t value)
     item->craftName[15] = str[1];
 }
 
-void ItemHelper::SetCraftStamina(CItem* item, uint16_t value)
+void ItemHelper::SetCraftStamina(CItem* item, uint value)
 {
     if (value > 99)
         value = 99;
@@ -166,7 +166,7 @@ void ItemHelper::InitCraftExpansion(CItem* item)
     item->craftAbsorption = 0;
 }
 
-void ItemHelper::SendDBAgentCraftName(CUser* user, CItem* item, uint8_t bag, uint8_t slot)
+void ItemHelper::SendDBAgentCraftName(CUser* user, CItem* item, uint bag, uint slot)
 {
     DBAgentItemCraftUpdateIncoming outgoing{};
     outgoing.billingId = user->billingId;
@@ -176,7 +176,7 @@ void ItemHelper::SendDBAgentCraftName(CUser* user, CItem* item, uint8_t bag, uin
     NetworkHelper::SendDBAgent(&outgoing, sizeof(DBAgentItemCraftUpdateIncoming));
 }
 
-void ItemHelper::SendDBAgentGems(CUser* user, CItem* item, uint8_t bag, uint8_t slot)
+void ItemHelper::SendDBAgentGems(CUser* user, CItem* item, uint bag, uint slot)
 {
     DBAgentItemGemUpdateIncoming outgoing{};
     outgoing.billingId = user->billingId;
@@ -187,7 +187,7 @@ void ItemHelper::SendDBAgentGems(CUser* user, CItem* item, uint8_t bag, uint8_t 
     NetworkHelper::SendDBAgent(&outgoing, sizeof(DBAgentItemGemUpdateIncoming));
 }
 
-void ItemHelper::SendDBAgentCloakInfo(CUser* user, CItem* item, uint8_t bag, uint8_t slot)
+void ItemHelper::SendDBAgentCloakInfo(CUser* user, CItem* item, uint bag, uint slot)
 {
     ItemHelper::SendDBAgentGems(user, item, bag, slot);
 }
